Remove dependency on `move-unit-test` from `sui-framework`
Move cost-calibration and unit testing code to `sui-move`.
