+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "crates/sui-faucet",
     "crates/sui-framework",
     "crates/sui-framework-build",
+    "crates/sui-framework-test",
     "crates/sui-indexer",
     "crates/sui-json",
     "crates/sui-json-rpc",
