+++ crates/sui-framework/Cargo.toml
@@ -36,7 +36,6 @@ move-compiler.workspace = true
 move-core-types.workspace = true
 move-package.workspace = true
 move-stdlib.workspace = true
-move-unit-test.workspace = true
 move-vm-runtime.workspace = true
 move-vm-test-utils.workspace = true
 move-vm-types.workspace = true
